Refactor for better readability, used experiment with functional programming

+++ CompanyInventory.Models/Company/NewCompanyRequest.cs
@@ -7,10 +7,10 @@ namespace CompanyInventory.Models.Company
     public class NewCompanyRequest
     {
         public string Name { get; set; }
-        public short FoundationYear { get; set; }
+        public int FoundationYear { get; set; }
         public List<EmployeeCompany> Employees { get; set; }
     }
-    
+
     public class NewCompanyRequestValidator : AbstractValidator<NewCompanyRequest>
     {
         public NewCompanyRequestValidator()
@@ -22,13 +22,13 @@ namespace CompanyInventory.Models.Company
             RuleFor(x => x.Name)
                 .MaximumLength(500)
                 .WithMessage("Company name can max 500 characters");
-            
+
             RuleFor(x => x.FoundationYear)
                 .GreaterThanOrEqualTo((short) 1000)
                 .WithMessage("Foundation year must be greater than 1000 year");
-            
+
             RuleFor(x => x.FoundationYear)
-                .LessThanOrEqualTo((short)9999)
+                .LessThanOrEqualTo((short) 9999)
                 .WithMessage("Foundation year must be less than 9999 year");
         }
     }
